@@ -1,41 +1,41 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-
-namespace blazor_wasm
+﻿namespace blazor_wasm
 {
     public partial class App
     {
-        List<Data> data = new List<Data>();
-        int selected;
-        int id = 1;
-        Random random = new Random(0);
+        public List<Data> Data { get; set; } = new();
+        public int Selected { get; set; }
 
-        string[] adjectives = new string[]
+        private readonly string[] adjectives = new string[]
         {
             "pretty", "large", "big", "small", "tall", "short", "long", "handsome", "plain", "quaint", "clean", "elegant", "easy", "angry", "crazy", "helpful", "mushy", "odd", "unsightly", "adorable", "important", "inexpensive", "cheap", "expensive", "fancy"
         };
-
-        string[] colours = new string[]
+        
+        private readonly string[] colours = new string[]
         {
             "red", "yellow", "blue", "green", "pink", "brown", "purple", "brown", "white", "black", "orange"
         };
-
-        string[] nouns = new string[]
+        
+        private readonly string[] nouns = new string[]
         {
             "table", "chair", "house", "bbq", "desk", "car", "pony", "cookie", "sandwich", "burger", "pizza", "mouse", "keyboard"
         };
 
+        private int _rowId = 1;
+
         List<Data> BuildData(int count = 1000)
         {
-            var result = new List<Data>();
+            List<Data> result = new(count);
             for (int i = 0; i < count; i++)
             {
-                result.Add(new Data
+                Data data = new()
                 {
-                    Id = this.id++,
-                    Label = adjectives[this.random.Next(adjectives.Length)] + " " + colours[this.random.Next(colours.Length)] + " " + nouns[this.random.Next(nouns.Length)]
-                });
+                    Id = _rowId++,
+                    Label = $"{adjectives[Random.Shared.Next(adjectives.Length)]} {colours[Random.Shared.Next(colours.Length)]} {nouns[Random.Shared.Next(nouns.Length)]}",
+                };
+
+                data.Select = (e) => Select(data);
+                data.Remove = (e) => Remove(data);
+                result.Add(data);
             }
 
             return result;
@@ -43,45 +43,45 @@ namespace blazor_wasm
 
         public void Select(Data item)
         {
-            this.selected = item.Id;
+            Selected = item.Id;
         }
 
-        void Delete(Data item)
+        void Remove(Data item)
         {
-            this.data.Remove(item);
+            Data.Remove(item);
         }
 
         void Run()
         {
-            this.data = this.BuildData();
+            Data = BuildData();
         }
         void Runlots()
         {
-            this.data = this.BuildData(10000);
+            Data = BuildData(10000);
         }
         void Add()
         {
-            this.data.AddRange(this.BuildData(1000));
+            Data.AddRange(BuildData(1000));
         }
         void Update()
         {
-            for (var i = 0; i < this.data.Count; i += 10)
+            for (var i = 0; i < Data.Count; i += 10)
             {
-                this.data[i].Label += " !!!";
+                Data[i].Label += " !!!";
             }
         }
         void Clear()
         {
-            this.data = new List<Data>();
-            this.selected = 0;
+            Data.Clear();
+            Selected = 0;
         }
         void SwapRows()
         {
-            if (this.data.Count > 998)
+            if (Data.Count > 998)
             {
-                var a = this.data[1];
-                this.data[1] = this.data[998];
-                this.data[998] = a;
+                var a = Data[1];
+                Data[1] = Data[998];
+                Data[998] = a;
             }
         }
     }
