@@ -1,16 +1,7 @@
-using System.Threading.Tasks;
+using blazor_wasm;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 
-namespace blazor_wasm
-{
-    public class Program
-    {
-        public static async Task Main(string[] args)
-        {
-            var builder = WebAssemblyHostBuilder.CreateDefault(args);
-            builder.RootComponents.Add<App>("app");
+var builder = WebAssemblyHostBuilder.CreateDefault(args);
+builder.RootComponents.Add<App>("app");
 
-            await builder.Build().RunAsync();
-        }
-    }
-}
+await builder.Build().RunAsync();
